pyproject.toml
[tool.poetry]
name = "superset-deploy"
version = "0.1.0"
description = "Apache Superset deployment on GCP with Pulumi"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "pulumi"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pulumi = "^3.0.0"
pulumi-gcp = "^7.0.0"
pulumi-docker = "^4.0.0"
pulumi-cloudflare = "^5.0.0"
pyyaml = "^6.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
requests = "^2.28.0"
click = "^8.0.0"
rich = "^13.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.20.0"
pytest-mock = "^3.10.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]